fix(luban): 同步上游 ReadFlong 符号扩展修复
同步 Unity 仓 Issue #57 上游修复（包仓 commit 1b05bcc）：
低 32 位先转 uint 再转 long，避免符号位污染高位。新增
long.MaxValue 与 0x123456789ABCDEF0 两个回归用例（还原
缺陷实现实证用例有效性后恢复修复版）。

+++ addons/com.gameframex.godot.luban/Runtime/ByteBuf.cs
@@ -847,7 +847,7 @@ namespace LuBan.Runtime
 #else
             int xl = (Bytes[ReaderIndex + 3] << 24) | ((Bytes[ReaderIndex + 2] << 16)) | (Bytes[ReaderIndex + 1] << 8) | (Bytes[ReaderIndex]);
             int xh = (Bytes[ReaderIndex + 7] << 24) | (Bytes[ReaderIndex + 6] << 16) | (Bytes[ReaderIndex + 5] << 8) | Bytes[ReaderIndex + 4];
-            x = ((long)xh << 32) | (long)xl;
+            x = ((long)xh << 32) | (long)(uint)xl;
 #endif
             ReaderIndex += 8;
             return x;
